+++ test/all.rb
@@ -19,6 +19,7 @@ class TestRunner
     load File.expand_path('cache_test.rb', __dir__)
     load File.expand_path('saytime_time_test.rb', __dir__)
     load File.expand_path('gps_test.rb', __dir__)
+    load File.expand_path('geocoding_config_test.rb', __dir__)
     puts "All #{@count} tests passed."
   end